Cambios

+++ safechat-core/src/main/java/com/pe/fisi/sw/safechat_core/modules/blacklist/model/Atacante.java
@@ -14,7 +14,7 @@ public class Atacante {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long idAtacante;
+    private Integer idAtacante;
 
     @Column(nullable = false, unique = true)
     private String valor; // correo, teléfono o dominio
