+++ safechat-core/src/main/java/com/pe/fisi/sw/safechat_core/modules/blacklist/repository/ListaNegraRepository.java
@@ -3,6 +3,6 @@ package com.pe.fisi.sw.safechat_core.modules.blacklist.repository;
 import com.pe.fisi.sw.safechat_core.modules.blacklist.model.ListaNegra;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ListaNegraRepository extends JpaRepository<ListaNegra, Long> {
-    boolean existsByUsuarioIdUsuarioAndAtacanteIdAtacante(Long idUsuario, Long idAtacante);
+public interface ListaNegraRepository extends JpaRepository<ListaNegra, Integer> {
+    boolean existsByUsuarioIdUsuarioAndAtacanteIdAtacante(Integer idUsuario, Integer idAtacante);
 }
