+++ safechat-core/src/main/java/com/pe/fisi/sw/safechat_core/modules/message/model/Mensaje.java
@@ -14,7 +14,7 @@ public class Mensaje {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_mensaje")
-    private Long idMensaje;
+    private Integer idMensaje;
 
     @ManyToOne
     @JoinColumn(name = "id_usuario", referencedColumnName = "id_usuario", nullable = true)
